Gathers sourcemaps tests
