+++ lib/tidy-function.test.js
@@ -1,5 +1,5 @@
 /* eslint-disable max-len */
-const run = require('../index.test.js');
+const { run } = require('../index.test.js');
 const { typical } = require('../test/sharedConfigs');
 
 /**
