+++ lib/tidy-shorthand-property.test.js
@@ -1,4 +1,4 @@
-const run = require('../index.test.js');
+const { run } = require('../index.test.js');
 const postcss = require('postcss');
 const tidyShorthandProperty = require('./tidy-shorthand-property');
