+++ lib/tidy-var.test.js
@@ -1,4 +1,4 @@
-const run = require('../index.test.js');
+const { run } = require('../index.test.js');
 const { typical, typicalWithBreakpoints } = require('../test/sharedConfigs');
 
 /**
